feat: converter schemas to ts
