+++ src/schemas/index.ts
@@ -1,0 +1,3 @@
+export { BoardSchema } from './board.schema';
+export { UserSchema } from './user.schema';
+export { TaskSchema } from './tasks.schema';
